+++ .gitignore
@@ -7,6 +7,7 @@
 /.gradle/
 /build/
 src/main/resources/docs/
+src/main/java/compile.bat
 
 # MacOS custom attributes files created by Finder
 .DS_Store
