--- src/main/java/Command.java
@@ -1,8 +1,0 @@
-public abstract class Command {
-    void execute(TaskList tasks, Ui ui) throws Exception {
-    }
-
-    boolean isExit() {
-        return false;
-    }
-}
