+++ src/main/java/command/Command.java
@@ -1,0 +1,12 @@
+package command;
+import duke.Ui;
+import task.TaskList;
+
+public abstract class Command {
+    public void execute(TaskList tasks, Ui ui) throws Exception {
+    }
+
+    public boolean isExit() {
+        return false;
+    }
+}
