+++ src/main/java/command/DeleteCommand.java
@@ -1,3 +1,8 @@
+package command;
+import duke.Ui;
+import task.Task;
+import task.TaskList;
+
 public class DeleteCommand extends Command {
     private int index;
