+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,7 @@
+package command;
+import duke.Ui;
+import task.TaskList;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui) {
