+++ src/main/java/command/UnMarkCommand.java
@@ -1,3 +1,8 @@
+package command;
+import duke.Ui;
+import task.Task;
+import task.TaskList;
+
 public class UnMarkCommand extends Command {
     private int index;
