+++ src/main/java/exception/InvalidArgumentException.java
@@ -1,3 +1,4 @@
+package exception;
 public class InvalidArgumentException extends Exception {
     @Override
     public String toString() {
