+++ src/main/java/exception/NoPeriodException.java
@@ -1,3 +1,4 @@
+package exception;
 public class NoPeriodException extends IncompleteCommandException {
 
     public NoPeriodException() {
