+++ src/main/java/task/Task.java
@@ -1,3 +1,4 @@
+package task;
 public class Task {
     private String description;
     private boolean isDone;
